Getting rid of some compiler warnings (G++ 4.4 and/or 4.5?).
git-svn-id: https://gxemul.svn.sourceforge.net/svnroot/gxemul/gxemul/trunk@5494 dbbdf1f2-f3a0-11dc-8807-f0e41a653009

+++ src/devices/bus_pci.cc
@@ -115,7 +115,7 @@ void bus_pci_data_access(struct cpu *cpu, struct pci_data *pci_data,
 	if (dev == NULL) {
 		if (writeflag == MEM_READ) {
 			if (pci_data->cur_reg == 0)
-				*data = -1;
+				*data = (uint64_t) -1;
 			else
 				*data = 0;
 		} else {
